+++ app/src/components/card/index.ts
@@ -1,1 +1,2 @@
 export * from "./Card";
+export * from "./types";
